+++ auction-server/src/auction/service/cancel_bid.rs
@@ -61,7 +61,7 @@ impl Service<Svm> {
         }
     }
 
-    #[tracing::instrument(skip_all, err)]
+    #[tracing::instrument(skip_all, err, fields(bid_id = %input.bid_id))]
     pub async fn cancel_bid(&self, input: CancelBidInput) -> Result<(), RestError> {
         // Lock the bid to prevent submission
         let bid_lock = self